rename qb to builder
