+++ src/qlient/__init__.py
@@ -7,6 +7,6 @@ from qlient.backend import Backend, HTTPBackend
 from qlient.client import Client
 from qlient.exceptions import QlientException
 from qlient.models import GraphQLResponse
-from qlient.qb import Fields
+from qlient.builder import Fields
 from qlient.schema.schema import Schema
 from qlient.settings import Settings
