+++ tests/test_builder_fields.py
@@ -1,11 +1,11 @@
 def test_fields_simple_single():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a")
     assert a.fields == ["a"]
 
 
 def test_fields_simple_multiple():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", "c")
     assert "a" in a.fields
     assert "b" in a.fields
@@ -13,14 +13,14 @@ def test_fields_simple_multiple():
 
 
 def test_fields_simple_multiple_duplicates():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", "c", "c", "c")
     assert len(a.fields) == 3
     assert a.fields.count("c") == 1
 
 
 def test_fields_simple_list():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields(["a", "b", "c"])
     assert "a" in a.fields
     assert "b" in a.fields
@@ -28,7 +28,7 @@ def test_fields_simple_list():
 
 
 def test_fields_simple_list_duplicates():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     fields = ["a", "b", "c", "c", "c"]
     a = Fields(fields)
     assert len(a.fields) == 3
@@ -36,7 +36,7 @@ def test_fields_simple_list_duplicates():
 
 
 def test_fields_complex_simple():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", c="d")
     assert "a" in a.fields
     assert "b" in a.fields
@@ -45,7 +45,7 @@ def test_fields_complex_simple():
 
 
 def test_fields_complex_simple_list():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", c=["a", "b"])
     assert "a" in a.fields
     assert "b" in a.fields
@@ -54,7 +54,7 @@ def test_fields_complex_simple_list():
 
 
 def test_fields_complex_nested_fields():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields(a=Fields("a"), b=Fields("b"))
     assert "a" in a.sub_fields
     assert a.sub_fields["a"].fields == ["a"]
@@ -64,35 +64,35 @@ def test_fields_complex_nested_fields():
 
 
 def test_fields_simple_eq_operator():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a")
     b = Fields("a")
     assert a == b
 
 
 def test_fields_simple_eq_operator_not():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a")
     b = Fields("b")
     assert a != b
 
 
 def test_fields_complex_eq_operator():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", b="c")
     b = Fields("a", b="c")
     assert a == b
 
 
 def test_fields_complex_eq_operator_not():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", b="c")
     b = Fields("a", b="d")
     assert a != b
 
 
 def test_fields_simple_add_operator_simple():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a")
     expected = Fields("a", "b")
     actual = a + "b"
@@ -100,7 +100,7 @@ def test_fields_simple_add_operator_simple():
 
 
 def test_fields_simple_add_operator_list():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a")
     expected = Fields("a", "b", "c")
     actual = a + ["b", "c"]
@@ -108,7 +108,7 @@ def test_fields_simple_add_operator_list():
 
 
 def test_fields_simple_add_operator_dict():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a")
     expected = Fields("a", b="c")
     actual = a + {"b": "c"}
@@ -116,7 +116,7 @@ def test_fields_simple_add_operator_dict():
 
 
 def test_fields_simple_add_operator_fields():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a")
     b = Fields("a", "b", c="d")
     expected = Fields("a", "b", c="d")
@@ -125,7 +125,7 @@ def test_fields_simple_add_operator_fields():
 
 
 def test_fields_complex_add_operator_simple():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", b="b")
     expected = Fields("a", "c", b="b")
     actual = a + "c"
@@ -133,7 +133,7 @@ def test_fields_complex_add_operator_simple():
 
 
 def test_fields_complex_add_operator_list():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", b="b")
     expected = Fields("a", "c", "z", b="b")
     actual = a + ["a", "c", "z"]
@@ -141,7 +141,7 @@ def test_fields_complex_add_operator_list():
 
 
 def test_fields_complex_add_operator_dict():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", b="b")
     expected = Fields("a", b=["b", "c", "e"])
     actual = a + {"b": ["c", "e"]}
@@ -149,7 +149,7 @@ def test_fields_complex_add_operator_dict():
 
 
 def test_fields_complex_add_operator_fields():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", b="b")
     b = Fields("a", "c")
     expected = Fields("a", "c", b="b")
@@ -158,7 +158,7 @@ def test_fields_complex_add_operator_fields():
 
 
 def test_fields_simple_sub_operator_simple():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b")
     expected = Fields("a")
     actual = a - "b"
@@ -166,7 +166,7 @@ def test_fields_simple_sub_operator_simple():
 
 
 def test_fields_simple_sub_operator_list():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", "c")
     expected = Fields("a")
     actual = a - ["b", "c"]
@@ -174,7 +174,7 @@ def test_fields_simple_sub_operator_list():
 
 
 def test_fields_simple_sub_operator_fields():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", "c")
     expected = Fields("a")
     actual = a - Fields("b", "c", "z")
@@ -182,7 +182,7 @@ def test_fields_simple_sub_operator_fields():
 
 
 def test_fields_complex_sub_operator_simple():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", c="c")
     expected = Fields("a", c="c")
     actual = a - "b"
@@ -190,7 +190,7 @@ def test_fields_complex_sub_operator_simple():
 
 
 def test_fields_complex_sub_operator_list():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", "z", c="c")
     expected = Fields("a", c="c")
     actual = a - ["b", "z"]
@@ -198,7 +198,7 @@ def test_fields_complex_sub_operator_list():
 
 
 def test_fields_complex_sub_operator_dict():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", "z", c=["a", "b", "c"])
     expected = Fields("a", "b", "z", c=["a"])
     actual = a - {"c": ["b", "c"]}
@@ -206,7 +206,7 @@ def test_fields_complex_sub_operator_dict():
 
 
 def test_fields_complex_sub_operator_fields():
-    from qlient.qb import Fields
+    from qlient.builder import Fields
     a = Fields("a", "b", "z", c="c")
     expected = Fields("a")
     actual = a - Fields("b", "z", c="c")
