+++ tests/test_builder_gql_query_builder.py
@@ -1,5 +1,5 @@
 def test_query_builder():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder()
     expected = "{ }"
     actual = builder.build()
@@ -7,7 +7,7 @@ def test_query_builder():
 
 
 def test_query_builder_operation_query():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder().operation("query")
     expected = "query { }"
     actual = builder.build()
@@ -15,7 +15,7 @@ def test_query_builder_operation_query():
 
 
 def test_query_builder_operation_query_name():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder().operation("query", name="MyQuery")
     expected = "query MyQuery { }"
     actual = builder.build()
@@ -23,7 +23,7 @@ def test_query_builder_operation_query_name():
 
 
 def test_query_builder_operation_query_name_variables():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder().operation("query", name="MyQueryVar", variables={"$foo": "String!"})
     expected = "query MyQueryVar($foo: String!) { }"
     actual = builder.build()
@@ -31,7 +31,7 @@ def test_query_builder_operation_query_name_variables():
 
 
 def test_query_builder_operation_mutation():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder().operation("mutation")
     expected = "mutation { }"
     actual = builder.build()
@@ -39,7 +39,7 @@ def test_query_builder_operation_mutation():
 
 
 def test_query_builder_operation_mutation_name():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder().operation("mutation", name="MyMutation")
     expected = "mutation MyMutation { }"
     actual = builder.build()
@@ -47,7 +47,7 @@ def test_query_builder_operation_mutation_name():
 
 
 def test_query_builder_operation_mutation_name_variables():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder().operation("mutation", name="MyMutationVar", variables={"$foo": "String!"})
     expected = "mutation MyMutationVar($foo: String!) { }"
     actual = builder.build()
@@ -55,7 +55,7 @@ def test_query_builder_operation_mutation_name_variables():
 
 
 def test_query_builder_action():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder().action("foo")
     expected = "{ foo }"
     actual = builder.build()
@@ -63,7 +63,7 @@ def test_query_builder_action():
 
 
 def test_query_builder_action_variables():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder().action("foo", variables={"bar": 1})
     expected = "{ foo(bar: 1) }"
     actual = builder.build()
@@ -71,7 +71,7 @@ def test_query_builder_action_variables():
 
 
 def test_query_builder_fields_simple():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder()
     builder = builder.action("get_person")
     builder = builder.fields("first_name", "last_name")
@@ -81,7 +81,7 @@ def test_query_builder_fields_simple():
 
 
 def test_query_builder_fields_nested():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder()
     builder = builder.action("get_person")
     builder = builder.fields("first_name", "last_name", hobby=["title", "description"])
@@ -91,7 +91,7 @@ def test_query_builder_fields_nested():
 
 
 def test_query_builder_full_query():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder()
     builder = builder.operation("query", name="PersonQuery", variables={"$search": "String"})
     builder = builder.action("get_person", variables={"search": "$search"})
@@ -102,7 +102,7 @@ def test_query_builder_full_query():
 
 
 def test_query_builder_full_mutation():
-    from qlient.qb import GQLQueryBuilder
+    from qlient.builder import GQLQueryBuilder
     builder = GQLQueryBuilder()
     builder = builder.operation("mutation", name="UpdatePerson", variables={"$id": "ID!"})
     builder = builder.action("update_person", variables={"id": "$id"})
